Split imageUploading into Vuex module
